# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Shop:
    properties:
      TIN:
        type: string
      about:
        type: string
      address:
        type: string
      city_id: {}
      company_logo:
        type: string
      company_name:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      legal_entity:
        type: string
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Supplier:
    properties:
      TIN:
        type: string
      about:
        type: string
      address:
        type: string
      city_id:
        type: integer
      company_logo:
        type: string
      company_name:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      legal_entity:
        type: string
      margin:
        type: integer
      minimum_order:
        type: integer
      name:
        type: string
      order_number:
        type: integer
      status:
        type: integer
      supply_dates:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      vacation_at:
        type: string
    type: object
...
  utils.DataResponse:
    properties:
      error_code:
        type: integer
      message: {}
    type: object
host: localhost:5005
info:
  contact: {}
  description: Api INSIGHT project
  title: Insight API SERVICE
  version: 1.0.0
paths:
...
  /shops/by-id:
    get:
      consumes:
      - application/json
      description: Просмотр магазина
      operationId: getShop
      parameters:
      - description: Введите данные
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getShop
      tags:
      - Shops
  /shops/edit:
    put:
      consumes:
      - application/json
      description: Редактирование магазина
      operationId: editShop
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Shop'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: editShop
      tags:
      - Shops
  /shops/list:
    get:
      consumes:
      - application/json
      description: Просмотр списока магазинов
      operationId: getAllShops
      parameters:
      - description: Введите данные
        in: query
        name: page
        required: true
        type: string
      - description: Введите данные
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getAllShops
      tags:
      - Shops
  /shops/new:
    post:
      consumes:
      - application/json
      description: Добавление нового магазина
      operationId: addNewShop
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Shop'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: addNewShop
      tags:
      - Shops
  /shops/rm:
    delete:
      consumes:
      - application/json
      description: Удаление магазина
      operationId: deleteShop
      parameters:
      - description: Введите данные
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: deleteShop
      tags:
      - Shops
  /suppliers/by-id:
    get:
      consumes:
      - application/json
      description: Просмотр поставщика
      operationId: getSupplier
      parameters:
      - description: Введите данные
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getSupplier
      tags:
      - Suppliers
  /suppliers/edit:
    put:
      consumes:
      - application/json
      description: Редактирование поставщика
      operationId: editSupplier
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Supplier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: editSupplier
      tags:
      - Suppliers
  /suppliers/list:
    get:
      consumes:
      - application/json
      description: Просмотр списока поставщиков
      operationId: getAllSuppliers
      parameters:
      - description: Введите данные
        in: query
        name: page
        required: true
        type: string
      - description: Введите данные
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: getAllSuppliers
      tags:
      - Suppliers
  /suppliers/new:
    post:
      consumes:
      - application/json
      description: Добавление нового поставщика
      operationId: addNewSupplier
      parameters:
      - description: Введите данные
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/models.Supplier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: addNewSupplier
      tags:
      - Suppliers
  /suppliers/rm:
    delete:
      consumes:
      - application/json
      description: Удаление поставщика
      operationId: deleteSupplier
      parameters:
      - description: Введите данные
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DataResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.DataResponse'
      security:
      - ApiKeyAuth: []
      summary: deleteSupplier
      tags:
      - Suppliers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"